Add dark mode to the app
