+++ frontend/src/App.tsx
@@ -1,0 +1,10 @@
+import { AppProvider } from '@/providers';
+import { AppRoutes } from '@/routing/routes';
+
+export function App() {
+  return (
+    <AppProvider>
+      <AppRoutes />
+    </AppProvider>
+  );
+}
