+++ frontend/src/providers/DarkModeProvider.tsx
@@ -1,0 +1,13 @@
+import { useColorMode } from '@chakra-ui/react';
+
+type Props = {
+  children: React.ReactNode;
+};
+
+export function DarkModeProvider({ children }: Props) {
+  const { setColorMode } = useColorMode();
+
+  setColorMode('dark');
+
+  return <>{children}</>;
+}
