+++ frontend/src/routing/routes/index.ts
@@ -1,5 +1,1 @@
-export { authRoute } from '../../features/auth/routes/auth.route';
-export { dashboardRoute } from './dashboard/dashboard.route';
-export { homeRoute } from './home/home.route';
-export { rootRoute } from './root/root.route';
-export { rootIndexRoute } from './root/root-index.route';
+export { AppRoutes } from './AppRoutes';
